+++ src/FileManager/drivers/FileMetadataGateway/FileMetadataModuleGateway.ts
@@ -1,18 +1,24 @@
 import { FileMetadataGateway } from '../../interfaces/FileMetadataGateway';
 import { UserToken } from '../../../shared/types';
-import { FileMetadataAdapter } from '../../../FileMetadata/adapters/FileMetadataAdapter';
 import { LearningObject } from '../../../shared/entity';
 import { FileMetadataFilter } from '../../../FileMetadata/typings';
+import { FileMetadata } from '../../../FileMetadata';
 
 export class FileMetadataModuleGateway extends FileMetadataGateway {
-    private adapter: FileMetadataAdapter = FileMetadataAdapter.getInstance();
 
-    getFileMeta(params: {
+    /**
+     * @inhertidoc
+     *
+     * Proxies FileMetadata module's `getFileMetadata`
+     * @returns {Promise<LearningObject.Material.File>}
+     * @memberof FileMetadataModuleGateway
+     */
+    getFileMetadata(params: {
         requester: UserToken;
         learningObjectId: string;
         id: string;
         filter: FileMetadataFilter;
     }): Promise<LearningObject.Material.File> {
-        return this.adapter.getFileMeta(params);
+        return FileMetadata.getFileMetadata(params);
     }
 }
