+++ src/FileManager/drivers/index.ts
@@ -1,0 +1,5 @@
+import { S3FileManager } from './FileManager/S3FileManager';
+import { ModuleLearningObjectGateway } from '../../FileMetadata/drivers';
+import { FileMetadataModuleGateway } from './FileMetadataGateway/FileMetadataModuleGateway';
+
+export {S3FileManager, ModuleLearningObjectGateway, FileMetadataModuleGateway}
