+++ src/FileManager/interfaces/index.ts
@@ -1,0 +1,5 @@
+import {FileManager} from './FileManager';
+import {FileMetadataGateway} from './FileMetadataGateway';
+import {LearningObjectGateway} from './LearningObjectGateway';
+
+export {FileManager, FileMetadataGateway, LearningObjectGateway}
