+++ src/LearningObjectSubmission/interactors/submitForReview.ts
@@ -3,7 +3,7 @@ import { Submission } from '../types/Submission';
 import { UserToken } from '../../shared/types';
 import { ResourceError, ResourceErrorReason } from '../../shared/errors';
 import { SubmissionDataStore } from '../SubmissionDatastore';
-import { LearningObjectAdapter } from '../../LearningObjects/LearningObjectAdapter';
+import { LearningObjectAdapter } from '../../LearningObjects/adapters/LearningObjectAdapter';
 import { EntityError } from '../../shared/entity/errors/entity-error';
 import { SubmissionPublisher } from './SubmissionPublisher';
