+++ src/LearningObjects/adapters/LearningObjectRouteHandler.ts
@@ -1,15 +1,13 @@
 import { Request, Response, Router } from 'express';
-import { mapErrorToResponseData, ResourceErrorReason } from '../shared/errors';
-import { DataStore } from '../shared/interfaces/DataStore';
+import { mapErrorToResponseData, ResourceErrorReason } from '../../shared/errors';
+import { DataStore } from '../../shared/interfaces/DataStore';
 import {
-  FileManager,
   LibraryCommunicator,
-} from '../shared/interfaces/interfaces';
-import { UserToken } from '../shared/types';
-import * as LearningObjectInteractor from './LearningObjectInteractor';
-import { LearningObject } from '../shared/entity';
-import { updateFileDescription, removeFile } from './LearningObjectInteractor';
-import { FileMeta, MaterialsFilter, LearningObjectFilter } from './typings';
+} from '../../shared/interfaces/interfaces';
+import { UserToken } from '../../shared/types';
+import * as LearningObjectInteractor from '../LearningObjectInteractor';
+import { LearningObject } from '../../shared/entity';
+import { FileMeta, MaterialsFilter, LearningObjectFilter } from '../typings';
 
 /**
  * Initializes an express router with endpoints for public Retrieving
@@ -60,7 +58,6 @@ export function initializePublic({
  * @export
  * @param {{
  *   dataStore: DataStore;
- *   fileManager: FileManager;
  *   library: LibraryCommunicator;
  * }} {
  *   dataStore,
@@ -72,12 +69,10 @@ export function initializePublic({
 export function initializePrivate({
   router,
   dataStore,
-  fileManager,
   library,
 }: {
   router: Router;
   dataStore: DataStore;
-  fileManager: FileManager;
   library: LibraryCommunicator;
 }) {
   const addLearningObject = async (req: Request, res: Response) => {
@@ -149,7 +144,6 @@ export function initializePrivate({
       const learningObjectName = req.params.learningObjectName;
       await LearningObjectInteractor.deleteLearningObject({
         dataStore,
-        fileManager,
         learningObjectName,
         library,
         user,
@@ -175,74 +169,6 @@ export function initializePrivate({
     }
   };
 
-  const addFileMeta = async (req: Request, res: Response) => {
-    try {
-      const requester: UserToken = req.user;
-      const authorUsername: string = req.params.username;
-      const learningObjectId: string = req.params.learningObjectId;
-      const fileMeta: FileMeta | FileMeta[] = req.body.fileMeta;
-      let fileMetaId;
-      if (Array.isArray(fileMeta)) {
-        fileMetaId = await LearningObjectInteractor.addLearningObjectFiles({
-          dataStore,
-          requester,
-          authorUsername,
-          learningObjectId,
-          fileMeta,
-        });
-      } else {
-        fileMetaId = await LearningObjectInteractor.addLearningObjectFile({
-          dataStore,
-          requester,
-          authorUsername,
-          learningObjectId,
-          fileMeta,
-        });
-      }
-      res.status(200).send({ fileMetaId });
-    } catch (e) {
-      const { code, message } = mapErrorToResponseData(e);
-      res.status(code).json({ message });
-    }
-  };
-
-  async function updateFileMetadata(req: Request, res: Response) {
-    try {
-      const objectId = req.params.learningObjectId;
-      const fileId = req.params.fileId;
-      const description = req.body.description;
-      await updateFileDescription({
-        fileId,
-        objectId,
-        description,
-        dataStore,
-      });
-      res.sendStatus(200);
-    } catch (e) {
-      const { code, message } = mapErrorToResponseData(e);
-      res.status(code).json({ message });
-    }
-  }
-
-  async function deleteFileMetadata(req: Request, res: Response) {
-    try {
-      const objectId = req.params.learningObjectId;
-      const fileId = req.params.fileId;
-      const username = req.user.username;
-      await removeFile({
-        dataStore,
-        fileManager,
-        objectId,
-        username,
-        fileId,
-      });
-      res.sendStatus(200);
-    } catch (e) {
-      const { code, message } = mapErrorToResponseData(e);
-      res.status(code).json({ message });
-    }
-  }
-
   router.route('/learning-objects').post(addLearningObject);
   router.patch('/learning-objects/:id', updateLearningObject);
   router.delete('/learning-objects/:learningObjectName', deleteLearningObject);
@@ -252,14 +178,4 @@ export function initializePrivate({
     '/learning-objects/:id/children/summary',
     getLearningObjectChildren,
   );
-  router.post(
-    '/users/:username/learning-objects/:learningObjectId/materials/files',
-    addFileMeta,
-  );
-  router
-    .route(
-      '/users/:username/learning-objects/:learningObjectId/materials/files/:fileId',
-    )
-    .patch(updateFileMetadata)
-    .delete(deleteFileMetadata);
 }
