+++ src/LearningObjects/gateways/index.ts
@@ -1,0 +1,4 @@
+import { ModuleFileMetadataGateway } from './FileMetadataGateway/ModuleFileMetadataGateway';
+import { ModuleFileManagerGateway } from './FileManagerGateway/ModuleFileManagerGateway';
+
+export { ModuleFileManagerGateway, ModuleFileMetadataGateway }
