+++ src/LearningObjects/interfaces/index.ts
@@ -1,0 +1,5 @@
+import { FileManagerGateway } from './FileManagerGateway';
+import { FileMetadataGateway } from './FileMetadataGateway';
+import { ReadMeBuilder } from './ReadMeBuilder';
+
+export {FileManagerGateway, FileMetadataGateway, ReadMeBuilder}
