+++ src/drivers/drivers.ts
@@ -3,12 +3,10 @@ import { ExpressRouteDriver } from './express/ExpressRouteDriver';
 import { ExpressAuthRouteDriver } from './express/ExpressAuthRouteDriver';
 import { ExpressAdminRouteDriver } from './express/ExpressAdminRouteDriver';
 import { MongoDriver } from './MongoDriver';
-import { S3Driver } from '../FileManager/drivers/FileManagerDriver/S3Driver';
 export {
   ExpressDriver,
   ExpressRouteDriver,
   ExpressAuthRouteDriver,
   ExpressAdminRouteDriver,
   MongoDriver,
-  S3Driver,
 };
