+++ src/drivers/express/ExpressAuthRouteDriver.ts
@@ -1,17 +1,13 @@
 import { Router } from 'express';
-import * as multer from 'multer';
 import { LearningObjectInteractor } from '../../interactors/interactors';
 import {
   DataStore,
-  FileManager,
   LibraryCommunicator,
 } from '../../shared/interfaces/interfaces';
 import {
   updateReadme,
-  removeFile,
-  updateFileDescription,
 } from '../../LearningObjects/LearningObjectInteractor';
-import * as LearningObjectRouteHandler from '../../LearningObjects/LearningObjectRouteHandler';
+import * as LearningObjectRouteHandler from '../../LearningObjects/adapters/LearningObjectRouteHandler';
 import * as LearningOutcomeRouteHandler from '../../LearningOutcomes/LearningOutcomeRouteHandler';
 import * as SubmissionRouteDriver from '../../LearningObjectSubmission';
 import * as ChangelogRouteHandler from '../../Changelogs/ChangelogRouteDriver';
@@ -19,20 +15,17 @@ import { reportError } from '../../shared/SentryConnector';
 import { UserToken } from '../../shared/types';
 import { ResourceErrorReason, mapErrorToResponseData } from '../../shared/errors';
 export class ExpressAuthRouteDriver {
-  private upload = multer({ storage: multer.memoryStorage() });
 
   constructor(
     private dataStore: DataStore,
-    private fileManager: FileManager,
     private library: LibraryCommunicator,
   ) { }
 
   public static buildRouter(
     dataStore: DataStore,
-    fileManager: FileManager,
     library: LibraryCommunicator,
   ): Router {
-    const e = new ExpressAuthRouteDriver(dataStore, fileManager, library);
+    const e = new ExpressAuthRouteDriver(dataStore, library);
     const router: Router = Router();
     e.setRoutes(router);
     return router;
@@ -78,7 +71,6 @@ export class ExpressAuthRouteDriver {
     LearningObjectRouteHandler.initializePrivate({
       router,
       dataStore: this.dataStore,
-      fileManager: this.fileManager,
       library: this.library,
     });
 
@@ -131,37 +123,12 @@ export class ExpressAuthRouteDriver {
         }
       },
     );
-    /**
-     * @deprecated
-     *
-     * TODO: Deprecate route in favor of more restful `/users/:username/learning-objects/:learningObjectId/materials/files/:fileId` route
-     */
-    router
-      .route('/files/:id/:fileId')
-      .patch(async (req, res) => {
-        try {
-          const objectId = req.params.id;
-          const fileId = req.params.fileId;
-          const description = req.body.description;
-          await updateFileDescription({
-            fileId,
-            objectId,
-            description,
-            dataStore: this.dataStore,
-          });
-          res.sendStatus(200);
-        } catch (e) {
-          const { code, message } = mapErrorToResponseData(e);
-          res.status(code).json({message});
-        }
-      });
     router.patch('/learning-objects/:id/pdf', async (req, res) => {
       try {
         const id = req.params.id;
         await updateReadme({
           id,
           dataStore: this.dataStore,
-          fileManager: this.fileManager,
         });
         res.sendStatus(200);
       } catch (e) {
@@ -214,7 +181,6 @@ export class ExpressAuthRouteDriver {
           const learningObjectNames = req.params.learningObjectNames.split(',');
           await LearningObjectInteractor.deleteMultipleLearningObjects({
             dataStore: this.dataStore,
-            fileManager: this.fileManager,
             library: this.library,
             user: req.user,
             learningObjectNames,
