+++ src/drivers/express/ExpressDriver.ts
@@ -2,7 +2,6 @@ import * as express from 'express';
 import * as bodyParser from 'body-parser';
 import {
   DataStore,
-  FileManager,
   LibraryCommunicator,
 } from '../../shared/interfaces/interfaces';
 import {
@@ -24,23 +23,20 @@ import {
 } from '../../shared/SentryConnector';
 import { LearningObjectSearch } from '../../LearningObjectSearch';
 import { FileMetadata } from '../../FileMetadata';
+import { FileManagerModule } from '../../FileManager/FileManagerModule';
 
 export class ExpressDriver {
   static app = express();
 
   static dataStore: DataStore;
 
-  static fileManager: FileManager;
-
   static library: LibraryCommunicator;
 
   static build(
     dataStore: DataStore,
-    fileManager: FileManager,
     library: LibraryCommunicator,
   ) {
     this.dataStore = dataStore;
-    this.fileManager = fileManager;
     this.library = library;
 
     this.attachConfigHandlers();
@@ -92,7 +88,6 @@ export class ExpressDriver {
       ExpressRouteDriver.buildRouter(
         this.dataStore,
         this.library,
-        this.fileManager,
       ),
     );
   }
@@ -103,10 +98,10 @@ export class ExpressDriver {
    */
   private static attachAuthenticatedRouters() {
     this.app.use(enforceAuthenticatedAccess);
+    this.app.use(FileManagerModule.expressRouter);
     this.app.use(
       ExpressAuthRouteDriver.buildRouter(
         this.dataStore,
-        this.fileManager,
         this.library,
       ),
     );
