+++ src/shared/interfaces/interfaces.ts
@@ -1,6 +1,5 @@
 
 import { DataStore } from './DataStore';
-import { FileManager } from '../../FileManager/interfaces/FileManager';
 import { LibraryCommunicator } from './LibraryCommunicator';
 
-export { DataStore, FileManager, LibraryCommunicator};
+export { DataStore, LibraryCommunicator};
