+++ src/shared/types/file-upload.ts
@@ -1,0 +1,7 @@
+import { Readable } from 'stream';
+
+export interface FileUpload {
+    path: string;
+    // FIXME: This should define the specific types it can take
+    data: Buffer|Uint8Array|Blob|string|Readable;
+}
